# PDFDataExtraction.WebAPI.Client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PDF Data Extraction",
    "version": "v1"
  },
  "paths": {
    "/api/PDFTextExtraction/detailed": {
      "post": {
        "tags": [
          "PDFTextExtraction"
        ],
        "summary": "Extract detailed text, including text position, size and font, from a PDF",
        "operationId": "DetailedTextExtraction",
        "parameters": [
          {
            "name": "wordLineDiff",
            "in": "query",
            "description": "Roughly the amount of pixels difference in the Y coordinates that is allowed for words on the same line. The default value is 10",
            "schema": {
              "type": "integer",
              "description": "Roughly the amount of pixels difference in the Y coordinates that is allowed for words on the same line. The default value is 10",
              "format": "int32"
            }
          },
          {
            "name": "whiteSpaceFactor",
            "in": "query",
            "description": "How wide the spacing between characters can be before the spacing is considered to be a whitespace, as a factor of the median character width.\r\nThe default value is 0.2",
            "schema": {
              "type": "number",
              "description": "How wide the spacing between characters can be before the spacing is considered to be a whitespace, as a factor of the median character width.\r\nThe default value is 0.2",
              "format": "double"
            }
          },
          {
            "name": "convertPdfToImages",
            "in": "query",
            "description": "Set this to `true` if the PDF should be converted to PNGs and included in the returned result. Default is `false`",
            "schema": {
              "type": "boolean",
              "description": "Set this to `true` if the PDF should be converted to PNGs and included in the returned result. Default is `false`",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "description": "The PDF to extract text from",
                    "format": "binary",
                    "nullable": true
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PDFTextExtractionResult"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PDFTextExtractionResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/PDFTextExtraction/simple": {
      "post": {
        "tags": [
          "PDFTextExtraction"
        ],
        "summary": "Extract simple text from a PDF",
        "operationId": "SimpleTextExtraction",
        "parameters": [
          {
            "name": "wordLineDiff",
            "in": "query",
            "description": "Roughly the amount of pixels difference in the Y coordinates that is allowed for words on the same line. The default value is 10",
            "schema": {
              "type": "integer",
              "description": "Roughly the amount of pixels difference in the Y coordinates that is allowed for words on the same line. The default value is 10",
              "format": "int32"
            }
          },
          {
            "name": "whiteSpaceFactor",
            "in": "query",
            "description": "How wide the spacing between characters can be before the spacing is considered to be a whitespace, as a factor of the median character width.\r\nThe default value is 0.2",
            "schema": {
              "type": "number",
              "description": "How wide the spacing between characters can be before the spacing is considered to be a whitespace, as a factor of the median character width.\r\nThe default value is 0.2",
              "format": "double"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "description": "The PDF to extract text from",
                    "format": "binary",
                    "nullable": true
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Point": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number",
            "format": "double"
          },
          "y": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "BoundingBox": {
        "type": "object",
        "properties": {
          "topLeft": {
            "$ref": "#/components/schemas/Point"
          },
          "bottomRight": {
            "$ref": "#/components/schemas/Point"
          }
        }
      },
      "Character": {
        "type": "object",
        "properties": {
          "bbox": {
            "$ref": "#/components/schemas/BoundingBox"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "font": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Word": {
        "type": "object",
        "properties": {
          "bbox": {
            "$ref": "#/components/schemas/BoundingBox"
          },
          "chars": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Character"
            },
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        }
      },
      "Line": {
        "type": "object",
        "properties": {
          "words": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Word"
            },
            "nullable": true
          },
          "lineNumberInPage": {
            "type": "integer",
            "format": "int32"
          },
          "lineNumberInDocument": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Page": {
        "type": "object",
        "properties": {
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Line"
            },
            "nullable": true
          },
          "width": {
            "type": "number",
            "format": "double"
          },
          "height": {
            "type": "number",
            "format": "double"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Document": {
        "type": "object",
        "properties": {
          "pages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Page"
            },
            "nullable": true
          }
        }
      },
      "PageAsImage": {
        "type": "object",
        "properties": {
          "contents": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "imageHeight": {
            "type": "integer",
            "format": "int32"
          },
          "imageWidth": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PDFFileMetadata": {
        "type": "object",
        "properties": {
          "fileMd5": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "textMd5": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PDFEmbeddedMetadata": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "author": {
            "type": "string",
            "nullable": true
          },
          "producer": {
            "type": "string",
            "nullable": true
          },
          "creator": {
            "type": "string",
            "nullable": true
          },
          "keywords": {
            "type": "string",
            "nullable": true
          },
          "creationDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modificationDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "PDFTextExtractionResult": {
        "type": "object",
        "properties": {
          "extractedData": {
            "$ref": "#/components/schemas/Document"
          },
          "pagesAsPNGs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PageAsImage"
            },
            "description": "The PDF file converted into PNGs, one PNG per page",
            "nullable": true
          },
          "fileMetaData": {
            "$ref": "#/components/schemas/PDFFileMetadata"
          },
          "pdfEmbeddedMetadata": {
            "$ref": "#/components/schemas/PDFEmbeddedMetadata"
          },
          "isSuccessful": {
            "type": "boolean",
            "description": "`true`, if the extraction was successful - `false` if it failed",
            "readOnly": true
          },
          "errorMessage": {
            "type": "string",
            "description": "If the extraction failed, this error message will describe what went wrong",
            "nullable": true
          }
        }
      }
    }
  }
}